// public/angular/app/interfaces/INotifCola.ts
export interface INotifCola {
    notificacion_cola: number;
    medio: number;
    tipo: number;
    fecha: string;
    persona: number;
    referencia: string;
    estado: number;
    fecha_desde: string;
    fecha_hasta: string;
    motivo_rechazo: number;
    usuario: number;
    observaciones: string;
    cola: number;
    notimasiva: number;
}